+++ PitchSpellingTools/PitchSpelling+EighthStepModifier.swift
@@ -19,7 +19,7 @@ extension PitchSpelling {
      - note: In 48-EDO, represents 1/8th-tone adjustment.
      May be useful for other cases (e.g., -14c adjustment for 5th partial, etc.).
      */
-    public enum EighthStepModifier: Float {
+    public enum EighthStepModifier: Double {
         
         /// None.
         case none = 0
